IRCTC-Microservices/IRCTC-Microservice-UserID/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asttokens==2.4.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.92.0
fastapi-cache2==0.2.1
frozenlist==1.3.3
geoip2==4.6.0
h11==0.14.0
httplib2==0.21.0
icecream==2.1.3
idna==3.4
ipinfo==4.4.2
maxminddb==2.2.0
motor==3.1.1
multidict==6.0.4
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
pendulum==2.1.2
pydantic==1.10.5
Pygments==2.17.2
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2022.7.1
pytzdata==2020.1
redis==5.0.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
termcolor==2.2.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
xlrd==2.0.1
XlsxWriter==3.0.8
yarl==1.8.2
